fix(docs-parsley): update algolia endpoints
